exa: factor out pools code to a standalone file
For consistency.
